+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 # Change Log
 
 ## [Unreleased]
+
+### Added
+
+- Announce walking onto hints
+
 ## [0.5.5] 2024-01-31
 
 ### Added